[refactor] 게시글 dto 이동
Motivation:
- dto 들은 항상 usecase 에 종속되기 때문에 dto 패키지에 있는 것 보다는 클래스 내부에 있는것이 더 좋을 것이다 생각해 이동하기로 결정했습니다.

Modification:

- `PostXXXDto.java` 패키지 이동
