+++ src/main/java/com/fasttime/domain/post/controller/PostController.java
@@ -3,12 +3,12 @@ package com.fasttime.domain.post.controller;
 import com.fasttime.domain.post.dto.controller.request.PostCreateRequestDto;
 import com.fasttime.domain.post.dto.controller.request.PostDeleteRequestDto;
 import com.fasttime.domain.post.dto.controller.request.PostUpdateRequestDto;
-import com.fasttime.domain.post.dto.service.request.PostCreateServiceDto;
-import com.fasttime.domain.post.dto.service.request.PostDeleteServiceDto;
-import com.fasttime.domain.post.dto.service.request.PostUpdateServiceDto;
 import com.fasttime.domain.post.dto.service.response.PostDetailResponseDto;
 import com.fasttime.domain.post.dto.service.response.PostsResponseDto;
 import com.fasttime.domain.post.service.PostCommandUseCase;
+import com.fasttime.domain.post.service.PostCommandUseCase.PostCreateServiceDto;
+import com.fasttime.domain.post.service.PostCommandUseCase.PostDeleteServiceDto;
+import com.fasttime.domain.post.service.PostCommandUseCase.PostUpdateServiceDto;
 import com.fasttime.domain.post.service.PostQueryUseCase;
 import com.fasttime.domain.post.service.PostQueryUseCase.PostSearchCondition;
 import com.fasttime.global.util.ResponseDTO;
