--- src/main/java/com/fasttime/domain/post/dto/service/request/PostDeleteServiceDto.java
@@ -1,18 +1,0 @@
-package com.fasttime.domain.post.dto.service.request;
-
-import java.time.LocalDateTime;
-import lombok.Getter;
-
-@Getter
-public class PostDeleteServiceDto {
-
-    private final Long postId;
-    private final Long memberId;
-    private final LocalDateTime deletedAt;
-
-    public PostDeleteServiceDto(Long postId, Long memberId, LocalDateTime deletedAt) {
-        this.postId = postId;
-        this.memberId = memberId;
-        this.deletedAt = deletedAt;
-    }
-}
