+++ src/main/java/com/fasttime/domain/post/service/PostCommandService.java
@@ -2,9 +2,6 @@ package com.fasttime.domain.post.service;
 
 import com.fasttime.domain.member.entity.Member;
 import com.fasttime.domain.member.service.MemberService;
-import com.fasttime.domain.post.dto.service.request.PostCreateServiceDto;
-import com.fasttime.domain.post.dto.service.request.PostDeleteServiceDto;
-import com.fasttime.domain.post.dto.service.request.PostUpdateServiceDto;
 import com.fasttime.domain.post.dto.service.response.PostDetailResponseDto;
 import com.fasttime.domain.post.entity.Post;
 import com.fasttime.domain.post.exception.NotPostWriterException;
