+++ src/test/java/com/fasttime/domain/member/unit/service/AdminServiceTest.java
@@ -4,13 +4,13 @@ import com.fasttime.domain.member.dto.MemberDto;
 import com.fasttime.domain.member.repository.MemberRepository;
 import com.fasttime.domain.member.service.AdminService;
 import com.fasttime.domain.member.service.MemberService;
-import com.fasttime.domain.post.dto.service.request.PostCreateServiceDto;
 import com.fasttime.domain.post.dto.service.response.PostDetailResponseDto;
 import com.fasttime.domain.post.dto.service.response.PostsResponseDto;
 import com.fasttime.domain.post.entity.Post;
 import com.fasttime.domain.post.entity.ReportStatus;
 import com.fasttime.domain.post.repository.PostRepository;
 import com.fasttime.domain.post.service.PostCommandService;
+import com.fasttime.domain.post.service.PostCommandUseCase.PostCreateServiceDto;
 import java.rmi.AccessException;
 import java.time.LocalDateTime;
 import java.util.List;
